+++ lib/src/constants/enums/enums_shelf.dart
@@ -1,2 +1,6 @@
 export 'auth_mode.dart';
+<<<<<<< HEAD
+=======
+export 'components.dart';
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
 export 'sign_up_modes.dart';