+++ lib/src/decorations/input_decorations.dart
@@ -31,7 +31,12 @@ class InputDeco {
     return InputDecoration(
       contentPadding: loginTheme.inputPadding ??
           EdgeInsets.symmetric(
+<<<<<<< HEAD
             vertical: _dynamicSize.height * (loginTheme.isLandscape ? 3.3 : 3),
+=======
+            vertical:
+                _dynamicSize.height * (loginTheme.isLandscape ? 3.3 : 2.6),
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
           ).copyWith(
             right: _dynamicSize.width *
                 (paddingFactor ?? (loginTheme.isLandscape ? 1 : 3)),