+++ lib/src/models/models_shelf.dart
@@ -1,3 +1,7 @@
+<<<<<<< HEAD
+=======
+export 'animated_component.dart';
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
 export 'animated_dialog_theme.dart';
 export 'language_dialog_theme.dart';
 export 'language_option.dart';