+++ lib/src/providers/login_theme.dart
@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
 
+<<<<<<< HEAD
 import '../models/animated_dialog_theme.dart';
 import 'login_view_theme.dart';
+=======
+import '../../animated_login.dart';
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
 
 /// [LoginTheme] is the provider for all design/theme related data.
 // ignore: prefer_mixin
@@ -192,12 +196,21 @@ class LoginTheme extends LoginViewTheme with ChangeNotifier {
   @override
   double? get spacingWithoutSocial => currentTheme.spacingWithoutSocial;
 
+<<<<<<< HEAD
+=======
+  @Deprecated('Use "actionButtonPadding" instead')
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   @override
   double? get spacingFormAndAction => currentTheme.spacingFormAndAction;
 
   @override
   EdgeInsets? get forgotPasswordPadding => currentTheme.forgotPasswordPadding;
 
+<<<<<<< HEAD
+=======
+  @Deprecated(
+      'Use "changeActionPadding" or "changeActionButtonPadding" instead')
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   @override
   double? get actionAndChangeActionSpacing =>
       currentTheme.actionAndChangeActionSpacing;
@@ -222,6 +235,41 @@ class LoginTheme extends LoginViewTheme with ChangeNotifier {
 
   @override
   Color? get loadingButtonColor => currentTheme.loadingButtonColor;
+<<<<<<< HEAD
   @override
   Color? get loadingSocialButtonColor => currentTheme.loadingSocialButtonColor;
+=======
+
+  @override
+  Color? get loadingSocialButtonColor => currentTheme.loadingSocialButtonColor;
+
+  @override
+  EdgeInsets? get titlePadding => currentTheme.titlePadding;
+
+  @override
+  EdgeInsets? get descriptionPadding => currentTheme.descriptionPadding;
+
+  @override
+  EdgeInsets? get changeActionPadding => currentTheme.changeActionPadding;
+
+  @override
+  EdgeInsets? get changeActionButtonPadding =>
+      currentTheme.changeActionButtonPadding;
+
+  @override
+  EdgeInsets? get useEmailPadding => currentTheme.useEmailPadding;
+
+  @override
+  EdgeInsets? get socialLoginPadding => currentTheme.socialLoginPadding;
+
+  @override
+  EdgeInsets? get formTitlePadding => currentTheme.formTitlePadding;
+
+  @override
+  EdgeInsets? get actionButtonPadding => currentTheme.actionButtonPadding;
+
+  @override
+  List<AnimatedComponent> get animatedComponentOrder =>
+      currentTheme.animatedComponentOrder;
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
 }