+++ lib/src/providers/login_view_theme.dart
@@ -1,6 +1,11 @@
 import 'package:flutter/material.dart';
 
+<<<<<<< HEAD
 import '../models/animated_dialog_theme.dart';
+=======
+import '../../animated_login.dart';
+import '../constants/enums/animation_type.dart';
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
 
 /// [LoginViewTheme] is the provider for all design/theme related data.
 class LoginViewTheme {
@@ -66,6 +71,42 @@ class LoginViewTheme {
     this.showLoadingSocialButton = true,
     this.loadingButtonColor,
     this.loadingSocialButtonColor,
+<<<<<<< HEAD
+=======
+    this.titlePadding,
+    this.descriptionPadding,
+    this.changeActionPadding,
+    this.changeActionButtonPadding,
+    this.useEmailPadding,
+    this.socialLoginPadding,
+    this.formTitlePadding,
+    this.actionButtonPadding,
+    this.animatedComponentOrder = const <AnimatedComponent>[
+      AnimatedComponent(
+          component: LoginComponents.logo, animationType: AnimationType.left),
+      AnimatedComponent(
+          component: LoginComponents.title, animationType: AnimationType.left),
+      AnimatedComponent(
+          component: LoginComponents.description,
+          animationType: AnimationType.left),
+      AnimatedComponent(
+          component: LoginComponents.formTitle,
+          animationType: AnimationType.left),
+      AnimatedComponent(component: LoginComponents.socialLogins),
+      AnimatedComponent(component: LoginComponents.useEmail),
+      AnimatedComponent(component: LoginComponents.form),
+      AnimatedComponent(component: LoginComponents.notHaveAnAccount),
+      AnimatedComponent(
+          component: LoginComponents.forgotPassword,
+          animationType: AnimationType.left),
+      AnimatedComponent(
+          component: LoginComponents.changeActionButton,
+          animationType: AnimationType.left),
+      AnimatedComponent(
+          component: LoginComponents.actionButton,
+          animationType: AnimationType.left),
+    ],
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   })  : assert(formWidthRatio >= 50, "Form width ratio should be at least 50."),
         assert(formElementsSpacing == null || formElementsSpacing <= 70,
             "Spacing between the form elements cannot be more than 70."),
@@ -188,7 +229,11 @@ class LoginViewTheme {
   /// Padding of the welcome part widget.
   final EdgeInsets? welcomePadding;
 
+<<<<<<< HEAD
   /// Padding of the form part widget.
+=======
+  /// Padding around the form elements.
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   final EdgeInsets? formPadding;
 
   /// Size of the logo in the welcome part.
@@ -251,6 +296,36 @@ class LoginViewTheme {
   /// Color of the loading indicator on a social login button.
   final Color? loadingSocialButtonColor;
 
+<<<<<<< HEAD
+=======
+  /// Custom padding around the title.
+  final EdgeInsets? titlePadding;
+
+  /// Custom padding around the description.
+  final EdgeInsets? descriptionPadding;
+
+  /// Padding around the change action CTA.
+  final EdgeInsets? changeActionPadding;
+
+  /// Padding around the change action button.
+  final EdgeInsets? changeActionButtonPadding;
+
+  /// Padding around the use email text.
+  final EdgeInsets? useEmailPadding;
+
+  /// Padding around the social login options.
+  final EdgeInsets? socialLoginPadding;
+
+  /// Padding around the form title.
+  final EdgeInsets? formTitlePadding;
+
+  /// Padding around the action button.
+  final EdgeInsets? actionButtonPadding;
+
+  /// Order of the screen components.
+  final List<AnimatedComponent> animatedComponentOrder;
+
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   /// Creates a copy login view theme with the given properties.
   LoginViewTheme copyWith({
     TextStyle? formTitleStyle,
@@ -311,6 +386,18 @@ class LoginViewTheme {
     bool? showLoadingSocialButton,
     Color? loadingButtonColor,
     Color? loadingSocialButtonColor,
+<<<<<<< HEAD
+=======
+    EdgeInsets? titlePadding,
+    EdgeInsets? descriptionPadding,
+    EdgeInsets? changeActionPadding,
+    EdgeInsets? changeActionButtonPadding,
+    EdgeInsets? useEmailPadding,
+    EdgeInsets? socialLoginPadding,
+    EdgeInsets? formTitlePadding,
+    EdgeInsets? actionButtonPadding,
+    List<AnimatedComponent>? animatedComponentOrder,
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   }) =>
       LoginViewTheme(
         formTitleStyle: formTitleStyle ?? this.formTitleStyle,
@@ -386,5 +473,19 @@ class LoginViewTheme {
         loadingButtonColor: loadingButtonColor ?? this.loadingButtonColor,
         loadingSocialButtonColor:
             loadingSocialButtonColor ?? this.loadingSocialButtonColor,
+<<<<<<< HEAD
+=======
+        titlePadding: titlePadding ?? this.titlePadding,
+        descriptionPadding: descriptionPadding ?? this.descriptionPadding,
+        changeActionPadding: changeActionPadding ?? this.changeActionPadding,
+        changeActionButtonPadding:
+            changeActionButtonPadding ?? this.changeActionButtonPadding,
+        useEmailPadding: useEmailPadding ?? this.useEmailPadding,
+        socialLoginPadding: socialLoginPadding ?? this.socialLoginPadding,
+        formTitlePadding: formTitlePadding ?? this.formTitlePadding,
+        actionButtonPadding: actionButtonPadding ?? this.actionButtonPadding,
+        animatedComponentOrder:
+            animatedComponentOrder ?? this.animatedComponentOrder,
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
       );
 }