+++ lib/src/widgets/form_part.dart
@@ -1,5 +1,6 @@
 part of '../../animated_login.dart';
 
+<<<<<<< HEAD
 class _FormPart extends StatefulWidget {
   /// Form part of the login screen.
   const _FormPart({
@@ -19,12 +20,20 @@ class _FormPart extends StatefulWidget {
     this.validateName = true,
     this.validateEmail = true,
     this.validatePassword = true,
+=======
+class _WebForm extends StatefulWidget {
+  /// Form part of the login screen.
+  const _WebForm({
+    required this.animationController,
+    required this.showForgotPassword,
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
     Key? key,
   }) : super(key: key);
 
   /// Main animation controller for the transition animation.
   final AnimationController animationController;
 
+<<<<<<< HEAD
   /// Form key in the form widget. It is especially used for input validations.
   final GlobalKey<FormState> formKey;
 
@@ -76,6 +85,16 @@ class _FormPart extends StatefulWidget {
 }
 
 class __FormPartState extends State<_FormPart> {
+=======
+  /// Indicates whether the forgot password option will be enabled.
+  final bool showForgotPassword;
+
+  @override
+  __WebFormState createState() => __WebFormState();
+}
+
+class __WebFormState extends State<_WebForm> {
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   /// Custom LoginTheme data, colors and styles on the screen.
   late LoginTheme loginTheme;
 
@@ -97,6 +116,7 @@ class __FormPartState extends State<_FormPart> {
   /// Animation that will change the location of the components of form.
   late final Animation<double> offsetAnimation;
 
+<<<<<<< HEAD
   /// Text Editing Controllers for the text form fields.
   late final TextEditingController _nameController;
   late final TextEditingController _emailController;
@@ -108,6 +128,8 @@ class __FormPartState extends State<_FormPart> {
 
   final FocusNode _confirmPasswordFocus = FocusNode();
 
+=======
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   late bool _isLandscape;
   late bool _isReverse;
 
@@ -116,11 +138,16 @@ class __FormPartState extends State<_FormPart> {
     super.initState();
 
     /// Initializes the appearance and disappearance animations
+<<<<<<< HEAD
     /// from outside to the screen for the form components.
+=======
+    /// from outside to the screen for the form LoginComponents.
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
     offsetAnimation = AnimationHelper(
       animationController: widget.animationController,
       animationCurve: context.read<LoginTheme>().animationCurve,
     ).tweenSequenceAnimation(80, 10);
+<<<<<<< HEAD
 
     final Auth readAuth = context.read<Auth>();
     _nameController =
@@ -131,12 +158,15 @@ class __FormPartState extends State<_FormPart> {
         TextEditingController(text: readAuth.password);
     _confirmPasswordController = widget.confirmPasswordController ??
         TextEditingController(text: readAuth.confirmPassword);
+=======
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   }
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
     final Auth auth = context.read<Auth>();
+<<<<<<< HEAD
     _emailController.value = TextEditingValue(text: auth.email ?? '');
   }
 
@@ -148,6 +178,13 @@ class __FormPartState extends State<_FormPart> {
     _confirmPasswordController.dispose();
     _confirmPasswordFocus.dispose();
     super.dispose();
+=======
+    auth
+      ..setEmailValue(auth.email)
+      ..setPasswordValue(auth.password)
+      ..setUsernameValue(auth.username)
+      ..setConfirmPasswordValue(auth.confirmPassword);
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   }
 
   @override
@@ -159,6 +196,7 @@ class __FormPartState extends State<_FormPart> {
     loginTexts = context.read<LoginTexts>();
     theme = Theme.of(context);
     auth = context.read<Auth>();
+<<<<<<< HEAD
     _fillFields();
     _initializeAnimations();
     return _isLandscape
@@ -172,6 +210,10 @@ class __FormPartState extends State<_FormPart> {
               child: child,
             ),
           );
+=======
+    _initializeAnimations();
+    return _webView;
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   }
 
   Widget get _webView => AnimatedBuilder(
@@ -197,6 +239,7 @@ class __FormPartState extends State<_FormPart> {
         ),
       );
 
+<<<<<<< HEAD
   Widget get _formColumn => Padding(
         padding: loginTheme.formPadding ??
             (_isLandscape
@@ -331,11 +374,211 @@ class __FormPartState extends State<_FormPart> {
 
   Form get _form => Form(
         key: _formKey,
+=======
+  Widget get _formColumn {
+    final List<Widget> items = <Widget>[];
+    for (final AnimatedComponent component
+        in loginTheme.animatedComponentOrder) {
+      items.addAll(_orderedComponent(component.component));
+    }
+    return Column(mainAxisAlignment: MainAxisAlignment.center, children: items);
+  }
+
+  List<Widget> _orderedComponent(LoginComponents component) {
+    switch (component) {
+      case LoginComponents.formTitle:
+        return <Widget>[if (_isLandscape) const _FormTitle()];
+      case LoginComponents.socialLogins:
+        if (auth.socialLogins != null && auth.socialLogins!.isNotEmpty) {
+          return <Widget>[const _SocialLoginOptions()];
+        }
+        return <Widget>[];
+      case LoginComponents.useEmail:
+        if (auth.socialLogins != null && auth.socialLogins!.isNotEmpty) {
+          return <Widget>[const _UseEmailText()];
+        }
+        return <Widget>[];
+      case LoginComponents.form:
+        return <Widget>[const _Form()];
+      case LoginComponents.forgotPassword:
+        return <Widget>[
+          if (_isReverse && widget.showForgotPassword) const _ForgotPassword()
+        ];
+      case LoginComponents.actionButton:
+        return <Widget>[const _ActionButton()];
+      default:
+        return <Widget>[Container()];
+    }
+  }
+
+  void _initializeAnimations() {
+    /// Initializes the transition animation from welcome part's width ratio
+    /// to 0 with custom animation curve and animation controller.
+    transitionAnimation = _isLandscape
+        ? Tween<double>(begin: 100 - loginTheme.formWidthRatio, end: 0).animate(
+            CurvedAnimation(
+              parent: widget.animationController,
+              curve: loginTheme.animationCurve,
+            ),
+          )
+        : AnimationHelper(
+            animationController: widget.animationController,
+            animationCurve: loginTheme.animationCurve,
+          ).tweenSequenceAnimation(120, 20);
+  }
+}
+
+class _ActionButton extends StatelessWidget {
+  const _ActionButton({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    final LoginTexts loginTexts = context.read<LoginTexts>();
+    final bool isReverse =
+        context.select<Auth, bool>((Auth auth) => auth.isReverse);
+    final bool isLandscape = loginTheme.isLandscape;
+    return Padding(
+      padding: loginTheme.actionButtonPadding ??
+          EdgeInsets.symmetric(
+              vertical: _customSpace(context, isReverse, isLandscape)),
+      child: RoundedButton(
+        buttonText: isReverse ? loginTexts.login : loginTexts.signUp,
+        onPressed: context.read<Auth>().action,
+        backgroundColor: isLandscape
+            ? Theme.of(context).primaryColor.withOpacity(.8)
+            : Colors.white,
+        buttonStyle: loginTheme.actionButtonStyle,
+      ),
+    );
+  }
+
+  double _customSpace(BuildContext context, bool isReverse, bool isLandscape) {
+    double factor = 2;
+    if (isLandscape) {
+      factor = 3;
+    }
+    return DynamicSize(context).height * factor;
+  }
+}
+
+class _FormTitle extends StatelessWidget {
+  const _FormTitle({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTexts loginTexts = context.read<LoginTexts>();
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    return Padding(
+      padding: loginTheme.formTitlePadding ?? EdgeInsets.zero,
+      child: BaseText(
+        context.select<Auth, bool>((Auth auth) => auth.isReverse)
+            ? loginTexts.loginFormTitle
+            : loginTexts.signUpFormTitle,
+        style: TextStyles(context)
+            .titleStyle(color: loginTheme.isLandscape ? null : Colors.white)
+            .merge(loginTheme.formTitleStyle),
+      ),
+    );
+  }
+}
+
+class _SocialLoginOptions extends StatelessWidget {
+  const _SocialLoginOptions({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    final DynamicSize dynamicSize = DynamicSize(context);
+    return Padding(
+      padding: loginTheme.socialLoginPadding ??
+          EdgeInsets.symmetric(vertical: dynamicSize.height * 1.4),
+      child: Wrap(
+        spacing: context.read<LoginTheme>().socialLoginsSpacing ??
+            dynamicSize.responsiveSize * 10,
+        alignment: WrapAlignment.center,
+        children: _socialLoginButtons(context, loginTheme.isLandscape),
+      ),
+    );
+  }
+
+  List<Widget> _socialLoginButtons(BuildContext context, bool isLandscape) {
+    final Auth auth = context.read<Auth>();
+    return List<Widget>.generate(
+      auth.socialLogins!.length,
+      (int index) => CircleWidget(
+        onTap: () async => auth.socialLoginCallback(index),
+        color: isLandscape ? null : Colors.white,
+        widthFactor: isLandscape ? 13 : 16,
+        child: Image.asset(auth.socialLogins![index].iconPath),
+      ),
+    );
+  }
+}
+
+class _UseEmailText extends StatelessWidget {
+  const _UseEmailText({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    final DynamicSize dynamicSize = DynamicSize(context);
+    return Padding(
+      padding: loginTheme.useEmailPadding ??
+          EdgeInsets.only(top: dynamicSize.height),
+      child: _useEmailText(context, loginTheme),
+    );
+  }
+
+  Widget _useEmailText(BuildContext context, LoginTheme loginTheme) {
+    final LoginTexts loginTexts = context.read<LoginTexts>();
+    return BaseText(
+      context.select<Auth, bool>((Auth auth) => auth.isReverse)
+          ? loginTexts.loginUseEmail
+          : loginTexts.signUpUseEmail,
+      style: TextStyles(context)
+          .subtitleTextStyle(
+              color: loginTheme.isLandscape ? Colors.black87 : Colors.white)
+          .merge(loginTheme.useEmailStyle),
+    );
+  }
+}
+
+class _Form extends StatefulWidget {
+  const _Form({Key? key}) : super(key: key);
+
+  @override
+  State<_Form> createState() => _FormState();
+}
+
+class _FormState extends State<_Form> {
+  final FocusNode _confirmPasswordFocus = FocusNode();
+
+  late Auth auth;
+
+  @override
+  void dispose() {
+    _confirmPasswordFocus.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    auth = context.read<Auth>();
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    final DynamicSize dynamicSize = DynamicSize(context);
+    return Padding(
+      padding: loginTheme.formPadding ??
+          EdgeInsets.symmetric(vertical: dynamicSize.height * 2),
+      child: Form(
+        key: auth.formKey,
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
         child: Wrap(
           direction: Axis.vertical,
           alignment: WrapAlignment.center,
           crossAxisAlignment: WrapCrossAlignment.center,
           spacing: context.read<LoginTheme>().formElementsSpacing ??
+<<<<<<< HEAD
               dynamicSize.height * (_isLandscape ? 2.2 : 1.5),
           children: _formElements,
         ),
@@ -464,4 +707,101 @@ class __FormPartState extends State<_FormPart> {
     _confirmPasswordController.value =
         TextEditingValue(text: auth.confirmPassword ?? '');
   }
+=======
+              dynamicSize.height * (loginTheme.isLandscape ? 2.2 : 1.2),
+          children: _formElements(auth, loginTheme),
+        ),
+      ),
+    );
+  }
+
+  List<Widget> _formElements(Auth auth, LoginTheme loginTheme) {
+    final LoginTexts loginTexts = context.read<LoginTexts>();
+    final bool isReverse =
+        context.select<Auth, bool>((Auth auth) => auth.isReverse);
+    return <Widget>[
+      if (!isReverse && auth.signUpMode != SignUpModes.confirmPassword)
+        CustomTextFormField(
+          controller: auth.nameController,
+          hintText: loginTexts.nameHint,
+          prefixIcon: Icons.person_outline,
+          prefixWidget: loginTheme.nameIcon,
+          validator: auth.nameValidator,
+          textInputAction: TextInputAction.next,
+          onChanged: auth.setUsername,
+          autofillHints: const <String>[
+            AutofillHints.username,
+            AutofillHints.newUsername,
+            AutofillHints.name,
+          ],
+          textInputType: TextInputType.name,
+        ),
+      CustomTextFormField(
+        controller: auth.emailController,
+        hintText: loginTexts.emailHint,
+        prefixIcon: Icons.email_outlined,
+        prefixWidget: loginTheme.emailIcon,
+        validator: auth.emailValidator,
+        textInputAction: TextInputAction.next,
+        onChanged: auth.setEmail,
+        autofillHints: const <String>[AutofillHints.email],
+        textInputType: TextInputType.emailAddress,
+      ),
+      ObscuredTextFormField(
+        controller: auth.passwordController,
+        hintText: loginTexts.passwordHint,
+        prefixIcon: Icons.password_outlined,
+        showPasswordVisibility: auth.showPasswordVisibility,
+        textInputAction:
+            auth.isSignup ? TextInputAction.next : TextInputAction.done,
+        onFieldSubmitted: (_) => auth.isSignup
+            ? _confirmPasswordFocus.requestFocus()
+            : auth.action(),
+        onChanged: auth.setPassword,
+        validator: auth.passwordValidator,
+      ),
+      if (!isReverse && auth.signUpMode != SignUpModes.name)
+        ObscuredTextFormField(
+          controller: auth.confirmPasswordController,
+          hintText: loginTexts.confirmPasswordHint,
+          prefixIcon: Icons.password_outlined,
+          showPasswordVisibility: auth.showPasswordVisibility,
+          onFieldSubmitted: (_) => auth.action(),
+          focusNode: _confirmPasswordFocus,
+          onChanged: auth.setConfirmPassword,
+          validator: auth.passwordValidator,
+        ),
+    ];
+  }
+}
+
+class _ForgotPassword extends StatelessWidget {
+  const _ForgotPassword({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    final Auth auth = context.read<Auth>();
+    final DynamicSize dynamicSize = DynamicSize(context);
+    final bool isLandscape = loginTheme.isLandscape;
+    return Container(
+      alignment: isLandscape ? Alignment.center : Alignment.topCenter,
+      padding: loginTheme.forgotPasswordPadding ??
+          (isLandscape
+              ? dynamicSize.lowTopPadding
+              : dynamicSize.lowBottomPadding),
+      child: BaseTextButton(
+        text: context.read<LoginTexts>().forgotPassword,
+        style: _defaultStyle(context, isLandscape)
+            .copyWith(decoration: TextDecoration.underline)
+            .merge(loginTheme.forgotPasswordStyle),
+        onPressed: () async => auth.onForgotPassword(auth.emailController.text),
+      ),
+    );
+  }
+
+  TextStyle _defaultStyle(BuildContext context, bool isLandscape) =>
+      TextStyles(context).subBodyStyle(
+          color: isLandscape ? Theme.of(context).primaryColor : Colors.white);
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
 }