+++ lib/src/widgets/welcome_components.dart
@@ -1,5 +1,6 @@
 part of '../../animated_login.dart';
 
+<<<<<<< HEAD
 class _LogoAndTexts extends StatelessWidget {
   /// Column of logo and welcome texts.
   const _LogoAndTexts({required this.logo, Key? key}) : super(key: key);
@@ -45,15 +46,78 @@ class _LogoAndTexts extends StatelessWidget {
   }
 
   Widget _logo(BuildContext context) {
+=======
+class _Title extends StatelessWidget {
+  /// Column of logo and welcome texts.
+  const _Title({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTexts loginTexts = context.read<LoginTexts>();
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    return Padding(
+      padding: loginTheme.titlePadding ??
+          EdgeInsets.symmetric(vertical: DynamicSize(context).height * .8),
+      child: BaseText(
+        context.select<Auth, bool>((Auth auth) => auth.isReverse)
+            ? loginTexts.welcomeBack
+            : loginTexts.welcome,
+        style: TextStyles(context)
+            .titleStyle(color: Colors.white)
+            .merge(context.read<LoginTheme>().welcomeTitleStyle),
+      ),
+    );
+  }
+}
+
+class _Description extends StatelessWidget {
+  const _Description({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final LoginTexts loginTexts = context.read<LoginTexts>();
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    return Padding(
+      padding: loginTheme.descriptionPadding ??
+          EdgeInsets.symmetric(
+              vertical: DynamicSize(context).height *
+                  (loginTheme.isLandscape ? 3 : 2)),
+      child: NotFittedText(
+        context.select<Auth, bool>((Auth auth) => auth.isReverse)
+            ? loginTexts.welcomeBackDescription
+            : loginTexts.welcomeDescription,
+        style: TextStyles(context)
+            .bodyStyle()
+            .merge(context.read<LoginTheme>().welcomeDescriptionStyle),
+      ),
+    );
+  }
+}
+
+class _Logo extends StatelessWidget {
+  const _Logo({required this.logo, Key? key}) : super(key: key);
+  final Widget? logo;
+
+  @override
+  Widget build(BuildContext context) {
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
     final DynamicSize dynamicSize = DynamicSize(context);
     final LoginTheme loginTheme = context.read<LoginTheme>();
     return Container(
       constraints: BoxConstraints.tight(loginTheme.logoSize ??
           Size.fromHeight(
+<<<<<<< HEAD
               dynamicSize.responsiveSize * (loginTheme.isLandscape ? 26 : 30))),
       padding: loginTheme.logoPadding ??
           EdgeInsets.only(
               bottom: dynamicSize.height * (loginTheme.isLandscape ? 4 : 2)),
+=======
+              dynamicSize.responsiveSize * (loginTheme.isLandscape ? 30 : 25))),
+      padding: loginTheme.logoPadding ??
+          EdgeInsets.symmetric(
+              vertical:
+                  dynamicSize.height * (loginTheme.isLandscape ? 2.8 : .8)),
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
       child: logo,
     );
   }
@@ -99,6 +163,7 @@ class _ChangeActionTitle extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final LoginTexts loginTexts = context.read<LoginTexts>();
+<<<<<<< HEAD
     return showButtonText
         ? Row(
             mainAxisAlignment: MainAxisAlignment.center,
@@ -108,6 +173,24 @@ class _ChangeActionTitle extends StatelessWidget {
             ],
           )
         : _changeActionTitle(context, loginTexts);
+=======
+    final LoginTheme loginTheme = context.watch<LoginTheme>();
+    return Padding(
+      padding: loginTheme.changeActionPadding ??
+          (loginTheme.isLandscape
+              ? EdgeInsets.symmetric(vertical: DynamicSize(context).height * 2)
+              : EdgeInsets.only(top: DynamicSize(context).height * 2)),
+      child: showButtonText
+          ? Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                _changeActionTitle(context, loginTexts),
+                _changeActionGesture(context, animate)
+              ],
+            )
+          : _changeActionTitle(context, loginTexts),
+    );
+>>>>>>> 15afffa4778b068a51a25368018e9755cabf80f5
   }
 
   Widget _changeActionTitle(BuildContext context, LoginTexts loginTexts) =>